+++ daemon/network/message_handler/log.rs
@@ -26,6 +26,7 @@ pub fn get_log(message: LogRequestMessage, state: &SharedState) -> Message {
                 {
                     Ok((stdout, stderr)) => (Some(stdout), Some(stderr)),
                     Err(err) => {
+                        // Fail early if there's some problem with getting the log output
                         return create_failure_message(format!(
                             "Failed reading process output file: {:?}",
                             err